Move to src layout
